Add initials extraction functions and corresponding tests
